creazione cartelle per esame

+++ Differenze finite/Differenze finite/ExportCSV.cpp
@@ -2,7 +2,7 @@
 
 
 
- void ExportCSV::exportfile(int tipo,vector<double> vec, int righe, System::String^ file,	 vector <double> Sexp,				//prendo i parametri in ingresso
+ void ExportCSV::exportfile(int tipo,vector<double> vec,size_t righe, System::String^ file,	 vector <double> Sexp,				//prendo i parametri in ingresso
 																							 vector <double> Kexp,
 																							 vector <double> rexp,
 																							 vector <double> sigmaexp,
