+++ Differenze finite/Differenze finite/PadreExport.cpp
@@ -1,7 +1,7 @@
 #include "PadreExport.h"
 
 
-void PadreExport::exportfile(int tipo, vector<double> vec, int righe, System::String^ file,     vector <double> Sexp,
+void PadreExport::exportfile(int tipo, vector<double> vec, size_t righe, System::String^ file,  vector <double> Sexp,
 																								vector <double> Kexp,
 																								vector <double> rexp,
 																								vector <double> sigmaexp,
