+++ Differenze finite/Differenze finite/PadreExport.h
@@ -7,7 +7,7 @@ ref class PadreExport			//creo la classe padre
 {
 public:
 	
-	virtual void exportfile(int tipo,vector<double> vec,int righe, System::String^ file,vector <double> Sexp,		// creo una funzione virtual per sfruttare il polimorfismo nelle classi figlie
+	virtual void exportfile(int tipo,vector<double> vec,size_t righe, System::String^ file,vector <double> Sexp,		// creo una funzione virtual per sfruttare il polimorfismo nelle classi figlie
 																						vector <double> Kexp,
 																						vector <double> rexp,
 																						vector <double> sigmaexp,
